s3writer/requirements.txt
git+https://github.com/owasp-sbot/OSBot-Utils.git
git+https://github.com/owasp-sbot/OSBot-Elastic.git
fastapi==0.64.0
fastapi-offline==1.2.2
python-dateutil
python-dotenv
python-json-logger
pydantic
requests
starlette == 0.13.6
urllib3
uvicorn
wheel
xmltodict
aiofiles

# for testing
coveralls
pytest
pytest-cov
pytest-datafiles
mock

# logging
elasticsearch==7.12.0

# needed for OSBOT-Elastic
git+https://github.com/owasp-sbot/OSBot-AWS.git
boto3

# system info
psutil

numpy
pandas